# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Q-Trust API
  version: "1.0.0"
  description: |
    Q-Trust is a post-quantum trust evaluation and attestation platform built on
    Base (Sepolia / Mainnet). The API provides:

    - **Assets**: register and verify on-chain CBOM (Cryptographic Bill of Materials) hashes.
    - **Orgs**: query organization summaries, assets, migrations, and audits.
    - **Attestations**: vendor product attestation via direct admin writes or
      EIP-712 gasless relay (no wallet needed).
    - **Migrations**: record and track post-quantum migration progress.
    - **Trust Evaluation**: policy-based PQC readiness scoring.
    - **Credentials**: placeholder VC issuance / verification (full Ed25519 VC stack in the Python SDK).
    - **Revocation / Policy / Schema / Trust-Anchor**: on-chain contract queries.
    - **Webhooks**: Redis-backed event subscriptions.
    - **Legacy**: pre-v1 alias routes (deprecated, sunset 2026-12-31).

    **Write routes** (`/v1/write/*`) require a valid `X-Api-Key` header in production.
    **Gasless relay routes** (`/v1/relay/*`) let vendors sign EIP-712 typed data off-chain;
    the relayer submits the transaction so vendors never need gas tokens.

  contact:
    name: Q-Trust Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001
    description: Local development server
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/write/assets:
    post:
      operationId: writeRegisterCBOM
      tags: [write]
      summary: Register a CBOM (admin)
      description: Admin-only. Registers a CBOM hash on-chain via the relayer.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCBOMRequest"
      responses:
        "200":
          description: CBOM registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResult"
        "400":
          description: Invalid cbomHash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/write/attestations:
    post:
      operationId: writeAttestProduct
      tags: [write]
      summary: Attest a product (admin)
      description: Admin-only. Records a product attestation on-chain via the relayer.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttestProductRequest"
      responses:
        "200":
          description: Attestation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResult"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Attestation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/write/migrations:
    post:
      operationId: writeRecordMigration
      tags: [write]
      summary: Record a migration (admin)
      description: Admin-only. Records a post-quantum migration on-chain via the relayer.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordMigrationRequest"
      responses:
        "200":
          description: Migration recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResult"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Migration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Relay (EIP-712 gasless) ────────────────────────────────────────
...
  /v1/evaluate:
    post:
      operationId: evaluateTrust
      tags: [evaluate]
      summary: Evaluate PQC readiness against a policy
      description: |
        Scores an organisation's post-quantum cryptography readiness by
        evaluating evidence against policy clauses with weighted satisfaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationResult"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Credentials ────────────────────────────────────────────────────
  /v1/credentials/issue:
    post:
      operationId: issueCredential
      tags: [credentials]
      summary: Issue a signed verifiable credential (Ed25519Signature2020)
      description: |
        API-key gated. Issues a W3C Verifiable Credential v2.0 signed with
        Ed25519Signature2020 under the backend's configured issuer key
        (`QTRUST_VC_ISSUER_KEY`; a deterministic dev key is generated when
        unset). The issuer is a self-contained `did:key` so the credential can
        be verified offline by `POST /v1/credentials/verify` or the Python
        SDK (`qtrust.vc.VCVerifier`).
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCredentialRequest"
      responses:
        "200":
          description: Credential issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialIssued"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/credentials/verify:
    post:
      operationId: verifyCredential
      tags: [credentials]
      summary: Verify a verifiable credential cryptographically (fail-closed)
      description: |
        Full cryptographic verification, fail-closed: the credential must be
        structurally valid, not expired, carry a proof, and the Ed25519
        signature must verify against the issuer DID's public key (did:key
        resolved locally, did:web via HTTPS with SSRF guard). Returns a
        structured result with per-check status. Byte-compatible with the
        Python SDK (`qtrust.vc.VCVerifier`) in both directions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCredentialRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialVerification"

  # ── Revocation ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Admin API key required for write and credential-issuance endpoints.
        Configured via `QTRUST_API_KEYS` (comma-separated). Enforced only in
        production when keys are configured.
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    RegisterCBOMRequest:
      type: object
      required: [cbomHash]
      properties:
        cbomHash:
          type: string
          pattern: "^0x[0-9a-fA-F]+$"
          description: 0x-prefixed CBOM hash
        metadataURI:
          type: string
          description: Optional metadata URI
          default: ""
    AttestProductRequest:
      type: object
      required: [productId, version, algorithm]
      properties:
        productId:
          type: string
        version:
          type: string
        algorithm:
          type: string
        supported:
          type: boolean
          default: false
        evidenceURI:
          type: string
          description: Optional evidence URI
          default: ""
    RecordMigrationRequest:
      type: object
      required: [migrationId, assetId, fromAlgorithm, toAlgorithm]
      properties:
        migrationId:
          type: string
        assetId:
          type: string
        fromAlgorithm:
          type: string
        toAlgorithm:
          type: string
        evidenceHash:
          type: string
          description: 0x-prefixed evidence hash (defaults to zero bytes32)
        evidenceURI:
          type: string
          description: Optional evidence URI
          default: ""
    TxResult:
      type: object
      properties:
        txHash:
          type: string
          description: On-chain transaction hash
        relayer:
          type: string
          description: Relayer address that submitted the transaction

    # ── Relay ──────────────────────────────────────────────────────
...
    EvaluateRequest:
      type: object
      required: [subject_did, policy_id, policy_version]
      properties:
        subject_did:
          type: string
          description: Organisation DID to evaluate
        policy_id:
          type: string
          description: Policy identifier
        policy_version:
          type: string
          description: Policy version
        evidence:
          type: array
          description: Optional evidence array (defaults to empty)
          items:
            type: object
            properties:
              evidence_id:
                type: string
              evidence_type:
                type: string
              claims:
                type: object
                additionalProperties: true
    EvaluationResult:
      type: object
      properties:
        assessment_id:
          type: string
          description: URN:uuid assessment identifier
        subject_did:
          type: string
        policy_id:
          type: string
        policy_version:
          type: string
        passed:
          type: boolean
          description: Whether all policy clauses are satisfied
        confidence:
          type: number
          format: float
          description: Weighted satisfaction score (0.0 - 1.0)
        evidence_used:
          type: array
          items:
            type: object
            properties:
              evidence_id:
                type: string
              evidence_type:
                type: string
              contribution:
                type: number
              matched_claims:
                type: array
                items:
                  type: string
        conflicts:
          type: array
          items: {}
        explanation:
          type: object
          additionalProperties:
            type: object
            properties:
              status:
                type: string
                enum: [satisfied, insufficient_evidence]
              evidence:
                type: string
              explanation:
                type: string
        evaluated_at:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time

    # ── Credentials ────────────────────────────────────────────────
    IssueCredentialRequest:
      type: object
      required: [subject_did]
      properties:
        schema_id:
          type: string
          description: Optional schema ID
        subject_did:
          type: string
          description: DID of the credential subject
        claims:
          type: object
          additionalProperties: true
          description: Optional claims payload
        expiration_date:
          type: string
          format: date
          description: Optional expiration date
    CredentialIssued:
      type: object
      properties:
        credential:
          type: object
          description: The full signed W3C Verifiable Credential (Ed25519Signature2020 proof attached)
        credential_id:
          type: string
          description: URN:uuid credential identifier
        issuer_did:
          type: string
          description: Backend issuer did:key
        subject_did:
          type: string
        schema_id:
          oneOf:
            - type: string
            - type: "null"
        issued_at:
          type: string
          format: date-time
        proof_type:
          type: string
          description: e.g. Ed25519Signature2020
        note:
          type: string
    VerifyCredentialRequest:
      type: object
      required: [presentation]
      properties:
        presentation:
          type: object
          description: Verifiable Credential object (with proof) to verify cryptographically
        verifier_did:
          type: string
          description: Optional verifier DID (reserved)
    CredentialVerification:
      type: object
      properties:
        valid:
          type: boolean
        reason:
          oneOf:
            - type: string
            - type: "null"
        detail:
          oneOf:
            - type: string
            - type: "null"
        issuer_did:
          oneOf:
            - type: string
            - type: "null"
        subject_did:
          oneOf:
            - type: string
            - type: "null"
        schema_id:
          oneOf:
            - type: string
            - type: "null"
        has_proof:
          oneOf:
            - type: boolean
            - type: "null"
        expired:
          oneOf:
            - type: boolean
            - type: "null"
        checked:
          type: object
          properties:
            structure:
              type: boolean
            expiration:
              type: boolean
            signature:
              type: boolean
        timestamp:
          type: string
          format: date-time
        verified_at:
          type: string
          format: date-time
        note:
          type: string
        error:
          type: string
          description: Present when validation fails

    # ── Revocation ─────────────────────────────────────────────────
...